Add support for separate Lexer and Parser
The peg subject is now made generic and now allows parsing of sequences
of other types, typically int. A new atom type is introduced for tokens
with the notation `[AAA]` where `AAA` can be any identifier available in
the scope of the parser (typically an enum). These changes allow NPeg to
parse a sequence of tokens that are collected by a lexer, and act on the
grammar of these tokens without having to (re)parse the string subject
to build an AST.

Thanks a lot to @sealmove for helping me think this through!